services/execution/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
celery==5.3.4
redis==5.0.1
pydantic==2.5.0
supabase==2.0.2
python-dotenv==1.0.0
httpx==0.25.2
prometheus-client==0.19.0